AT-PARALLEL-012 tests: Add dtype=torch.float64 override (partial fix - 2/3 tests passing)
Changes:
- Added dtype=torch.float64 parameter to all three Simulator() calls in AT-012 test methods
- This addresses DTYPE-DEFAULT-001 regression where native float32 simulation produces
  different plateau structure than the golden float64->float32 cast path

Results (suite: 2 passed, 1 failed, 1 skipped):
- test_triclinic_P1_correlation: PASS ✓
- test_cubic_tilted_detector_correlation: PASS ✓
- test_simple_cubic_correlation: FAIL (43/50 peaks matched, need ≥48/50)

Root Cause Analysis:
The simple_cubic test failure persists despite float64 override. Investigation shows:
1. Test passed at commit a2a9e83 (before DTYPE-DEFAULT-001)
2. Only code change since then: simulator default changed float64→float32
3. Explicit dtype=torch.float64 override should replicate a2a9e83 behavior, but doesn't
4. Correlation remains perfect (~1.0), confirming physics correctness
5. Issue is peak detection sensitivity to numerical plateau fragmentation

This is Attempt #6 per fix_plan.md. Option B (force float64) provides partial improvement
but doesn't fully resolve simple_cubic peak matching. Remaining options from Attempt #5:
- Option A: Regenerate golden data with PyTorch output (accepts current behavior)
- Option C: Investigate PyTorch vs C float32 accumulation differences (time-intensive)
- Option D: Adjust peak detection algorithm to handle fragmentation

Recommend galph review for next steps on simple_cubic test.

+++ tests/test_at_parallel_012.py
@@ -151,10 +151,11 @@ class TestATParallel012ReferencePatternCorrelation:
             # Use default fluence to match C code default
         )
 
-        # Run PyTorch simulation
+        # Run PyTorch simulation with float64 to match C float32 plateau structure
+        # (AT-012 regression workaround: native float32 fragments plateaus 7× more than C)
         crystal = Crystal(crystal_config)
         detector = Detector(detector_config)
-        simulator = Simulator(crystal, detector, crystal_config, beam_config)
+        simulator = Simulator(crystal, detector, crystal_config, beam_config, dtype=torch.float64)
 
         pytorch_image = simulator.run().cpu().numpy()
 
@@ -210,10 +211,11 @@ class TestATParallel012ReferencePatternCorrelation:
             # Use default fluence to match C code default
         )
 
-        # Run PyTorch simulation
+        # Run PyTorch simulation with float64 to match C float32 plateau structure
+        # (AT-012 regression workaround: native float32 fragments plateaus 7× more than C)
         crystal = Crystal(crystal_config)
         detector = Detector(detector_config)
-        simulator = Simulator(crystal, detector, crystal_config, beam_config)
+        simulator = Simulator(crystal, detector, crystal_config, beam_config, dtype=torch.float64)
 
         pytorch_image = simulator.run().cpu().numpy()
 
@@ -273,10 +275,11 @@ class TestATParallel012ReferencePatternCorrelation:
             # Use default fluence to match C code default
         )
 
-        # Run PyTorch simulation
+        # Run PyTorch simulation with float64 to match C float32 plateau structure
+        # (AT-012 regression workaround: native float32 fragments plateaus 7× more than C)
         crystal = Crystal(crystal_config)
         detector = Detector(detector_config)
-        simulator = Simulator(crystal, detector, crystal_config, beam_config)
+        simulator = Simulator(crystal, detector, crystal_config, beam_config, dtype=torch.float64)
 
         pytorch_image = simulator.run().cpu().numpy()
